backend/requirements.txt
asgiref==3.11.1
django==5.2.11
django-admin-sortable2==2.3.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==25.1.0
markdown==3.10.2
packaging==26.0
pyjwt==2.11.0
python-frontmatter==1.1.0
pyyaml==6.0.3
sqlparse==0.5.5
